requirements.txt
# AI Framework (Open Source)
crewai>=0.28.8,<0.30.0
crewai-tools>=0.1.6,<0.3.0

# LLM Support (NO PAID APIs - Uses local Ollama)
httpx<0.28.0
langchain>=0.1.0,<0.2.0
langchain-core>=0.1.0,<0.2.0
pydantic>=2.6.0,<2.8.0

# Configuration
python-dotenv>=1.0.0

# Web Scraping & Data Collection (Open Source)
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
feedparser>=6.0.0

# Data Processing & Analysis (Open Source)
sentence-transformers>=2.2.0
scikit-learn>=1.3.0
numpy>=1.26.0
pandas>=2.1.0

# Utilities
setuptools>=69.0.0,<80.0.0

# NOTE: No groq or paid API packages
# Uses Ollama for LOCAL, FREE, OPEN-SOURCE LLM
